requirements.txt
Django==4.2.7
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
tensorflow-cpu==2.14.0
gunicorn==20.1.0